requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
compressai==1.2.6
contourpy==1.3.0
cycler==0.12.1
DISTS-pytorch==0.1
einops==0.8.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.4
kiwisolver==1.4.7
lpips==0.1.4
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
psutil==6.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-msssim==1.0.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scipy==1.13.1
six==1.16.0
sympy==1.13.3
thop==0.1.0
torch==2.1.0
torch-geometric==2.6.1
torchaudio==2.4.1
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
zipp==3.21.0
